+++ setup/about/about.php
@@ -7,7 +7,7 @@
  * @copyright   2014 WebMan - Oliver Juhas
  *
  * @since    1.0
- * @version  1.9.4
+ * @version  1.9.5
  *
  * CONTENT:
  *  1) Requirements check
@@ -150,7 +150,7 @@
 	 * Render the "About" screen content
 	 *
 	 * @since    1.0
-	 * @version  1.9.3
+	 * @version  1.9.5
 	 */
 	if ( ! function_exists( 'wm_about_screen' ) ) {
 		function wm_about_screen() {
@@ -283,38 +283,6 @@
 									<a href="http://codex.wordpress.org/Editing_wp-config.php#WordPress_Upgrade_Constants" target="_blank"><?php esc_html_e( 'In that case please read the instructions &raquo;', 'mustang-lite' ); ?></a>
 								</em>
 							</p>
-
-						<!-- Special note -->
-
-							<div class="wm-notes special">
-
-								<h2 class="mt0"><strong><?php esc_html_e( 'Installing the theme demo content', 'mustang-lite' ); ?></strong></h2>
-
-								<p>
-									<?php esc_html_e( 'You can install the theme demo content including pages, posts, custom post types, layouts, menus and widgets directly from your WordPress dashboard by clicking the button bellow.', 'mustang-lite' ); ?>
-								</p>
-
-								<p>
-									<?php esc_html_e( 'Alternatively (such as when the automated installation fails) you can follow theme documentation instructions for manual demo content installation.', 'mustang-lite' ); ?>
-									<a href="https://webmandesign.github.io/docs/mustang/#demo-content" target="_blank"><?php esc_html_e( 'Read the instructions &raquo;', 'mustang-lite' ); ?></a>
-								</p>
-
-								<?php if ( ! ( class_exists( 'OCDI_Plugin' ) || class_exists( 'PT_One_Click_Demo_Import' ) ) ) : ?>
-
-									<a href="<?php echo esc_url( admin_url( 'themes.php?page=tgmpa-install-plugins' ) ); ?>" class="button button-hero"><strong><?php esc_html_e( 'Install and run "One Click Demo Import" plugin', 'mustang-lite' ); ?></strong></a>
-
-								<?php else : ?>
-
-									<a href="<?php echo esc_url( 'themes.php?page=pt-one-click-demo-import' ); ?>" class="button button-hero button-primary"><strong><?php esc_html_e( 'Install theme demo content', 'mustang-lite' ); ?></strong></a>
-
-									<br>
-									<small><em>
-										<?php esc_html_e( 'Or head over to Appearance &raquo; Import Demo Data to start the import process.', 'mustang-lite' ); ?>
-									</em></small>
-
-								<?php endif; ?>
-
-							</div>
 
 						</div>